+++ src/exceptions.py
@@ -4,6 +4,17 @@ class PluginError(Exception):
     pass
 
 
-class ConfigurationError(PluginError):
-    """Errors in runtime configuration"""
+class InvalidCredentialsError(PluginError):
+    """Errors in polarion credentials"""
+    pass
+
+
+class TestCaseTypeError(PluginError):
+    '''Error in Test Case result update 
+    in case the Test Type is not configured as Automated'''
+    pass
+
+class TestCaseNotFoundError(PluginError):
+    '''Error in the validation of the test cases needed 
+    for update on Polarion.'''
     pass