+++ src/runtime_settings.py
@@ -5,11 +5,26 @@ class Settings:
     POLARION_HOST = ''
     POLARION_USER = ''
     POLARION_PASSWORD = ''
-    # POLARION_TOKEN = ''
+    POLARION_TOKEN = ''
     POLARION_PROJECT_ID = ''
     POLARION_TEST_RUN = ''
+    WEB_URL = ''
+    ALLUREDIR = ''
+    ALLOW_COMMENTS = None
 
 
 class TestExecutionResult:
-    polarion_tags = {}
-    polarion_id_test_result = {}
+    polarion_test_mapping = {}  # test_name: polarion_test_id (str: str)
+
+    result_polarion_mapping = {}  # polarion_test_id: test_results (str: list)
+    test_polarion_mapping = {}  # Inverted polarion_test_mapping dict
+
+    uid_test_mapping = {}  # test_name: uid (str: str)
+    parent_uid = None
+
+
+class PolarionTestRunRefs:
+    client = None
+    project = None
+    run = None
+    test_cases = list()